@@ -2,8 +2,8 @@
     "name": "zfegg/expressive-test",
     "description": "Zend Expressive abstract test case for PHPUnit",
     "require": {
-        "php": ">=7.2",
-        "phpunit/phpunit": "^6.0 | ^7.0 | ^8.0 | ^9.0",
+        "php": ">=8.1",
+        "phpunit/phpunit": ">=8.0",
         "mezzio/mezzio": "^3.0",
         "laminas/laminas-diactoros": "^3.0",
         "laminas/laminas-servicemanager": "^3.3",
